+++ data-raw/sysdata.R
@@ -6,6 +6,13 @@ plot_meta <- list(
   ),
   plot_count_by_date = list(
     x = "Date"
+  ),
+  plot_stacked_bar = list(
+    x = "character_or_factor"
+  ),
+  plot_cat_count_by_year = list(
+    x = "Date",
+    y = "character_or_factor"
   )
 )